RF: Initial refactor of the Camera sub-package
