+++ psychopy/hardware/camera/__init__.py
@@ -55,15 +55,24 @@ from psychopy.visual.movies.frame import MovieFrame, NULL_MOVIE_FRAME_INFO
 from psychopy.sound.microphone import Microphone
 import psychopy.logging as logging
 from ffpyplayer.player import MediaPlayer
-from ffpyplayer.writer import MediaWriter
-from ffpyplayer.pic import SWScale
 from ffpyplayer.tools import list_dshow_devices, get_format_codec
-# Something in moviepy.editor's initialisation breaks Mouse, so import these
-# from the source instead
-# from moviepy.editor import VideoFileClip, AudioFileClip, CompositeAudioClip
-from moviepy.video.io.VideoFileClip import VideoFileClip
-from moviepy.audio.io.AudioFileClip import AudioFileClip
-from moviepy.audio.AudioClip import CompositeAudioClip
+
+# relative imports from `Camera`
+from .constants import (
+    VIDEO_DEVICE_ROOT_LINUX,
+    CAMERA_UNKNOWN_VALUE,
+    CAMERA_NULL_VALUE,
+    CAMERA_TEMP_FILE_VIDEO,
+    CAMERA_TEMP_FILE_AUDIO,
+    CAMERA_API_AVFOUNDATION,
+    CAMERA_API_DIRECTSHOW,
+    CAMERA_API_UNKNOWN,
+    CAMERA_API_NULL,
+    CAMERA_FRAMERATE_NTSC,
+    CAMERA_FRAMERATE_NOMINAL_NTSC)
+from .info import CameraInfo, getCameras, getCameraDescriptions
+from .stream import StreamData, StreamStatus, StreamWriterThread, renderVideo
+
 import uuid
 import threading
 import queue
@@ -74,34 +83,6 @@ import time
 # ------------------------------------------------------------------------------
 # Constants
 #
-
-VIDEO_DEVICE_ROOT_LINUX = '/dev'
-CAMERA_UNKNOWN_VALUE = u'Unknown'  # fields where we couldn't get a value
-CAMERA_NULL_VALUE = u'Null'  # fields where we couldn't get a value
-# camera operating modes
-# CAMERA_MODE_VIDEO = u'video'
-# CAMERA_MODE_CV = u'cv'
-# CAMERA_MODE_PHOTO = u'photo'
-# default names for video and audio tracks in the temp directory
-CAMERA_TEMP_FILE_VIDEO = u'video.mp4'
-CAMERA_TEMP_FILE_AUDIO = u'audio.wav'
-
-# camera API flags, these specify which API camera settings were queried with
-CAMERA_API_AVFOUNDATION = u'AVFoundation'  # mac
-CAMERA_API_DIRECTSHOW = u'DirectShow'      # windows
-# CAMERA_API_VIDEO4LINUX = u'Video4Linux'    # linux
-# CAMERA_API_OPENCV = u'OpenCV'              # opencv, cross-platform API
-CAMERA_API_UNKNOWN = u'Unknown'            # unknown API
-CAMERA_API_NULL = u'Null'                  # empty field
-
-# camera libraries for playback nad recording
-CAMERA_LIB_FFPYPLAYER = u'FFPyPlayer'
-CAMERA_LIB_UNKNOWN = u'Unknown'
-CAMERA_LIB_NULL = u'Null'
-
-# special values
-CAMERA_FRAMERATE_NOMINAL_NTSC = '30.000030'
-CAMERA_FRAMERATE_NTSC = 30.000030
 
 # FourCC and pixel format mappings, mostly used with AVFoundation to determine
 # the FFMPEG decoder which is most suitable for it. Please expand this if you
@@ -155,532 +136,6 @@ class PlayerNotAvailableError(Exception):
 # ------------------------------------------------------------------------------
 # Classes
 #
-
-class CameraInfo:
-    """Information about a specific operating mode for a camera attached to the
-    system.
-
-    Parameters
-    ----------
-    name : str
-        Camera name retrieved by the OS. This may be a human-readable name
-        (i.e. DirectShow on Windows), an index on MacOS or a path (e.g.,
-        `/dev/video0` on Linux).
-    frameSize : ArrayLike
-        Resolution of the frame `(w, h)` in pixels.
-    frameRate : ArrayLike
-        Allowable framerate for this camera mode.
-    pixelFormat : str
-        Pixel format for the stream. If `u'Null'`, then `codecFormat` is being
-        used to configure the camera.
-    codecFormat : str
-        Codec format for the stream.  If `u'Null'`, then `pixelFormat` is being
-        used to configure the camera. Usually this value is used for high-def
-        stream formats.
-
-    """
-    __slots__ = [
-        '_index',
-        '_name',
-        '_frameSize',
-        '_frameRate',
-        '_pixelFormat',
-        '_codecFormat',
-        '_cameraLib',
-        '_cameraAPI'  # API in use, e.g. DirectShow on Windows
-    ]
-
-    def __init__(self,
-                 index=-1,
-                 name=CAMERA_NULL_VALUE,
-                 frameSize=(-1, -1),
-                 frameRate=(-1, -1),
-                 pixelFormat=CAMERA_UNKNOWN_VALUE,
-                 codecFormat=CAMERA_UNKNOWN_VALUE,
-                 cameraLib=CAMERA_NULL_VALUE,
-                 cameraAPI=CAMERA_API_NULL):
-
-        self.index = index
-        self.name = name
-        self.frameSize = frameSize
-        self.frameRate = frameRate
-        self.pixelFormat = pixelFormat
-        self.codecFormat = codecFormat
-        self.cameraLib = cameraLib
-        self.cameraAPI = cameraAPI
-
-    def __repr__(self):
-        return (f"CameraInfo(index={repr(self.index)}, "
-                f"name={repr(self.name)}, "
-                f"frameSize={repr(self.frameSize)}, "
-                f"frameRate={self.frameRate}, "
-                f"pixelFormat={repr(self.pixelFormat)}, "
-                f"codecFormat={repr(self.codecFormat)}, "
-                f"cameraLib={repr(self.cameraLib)}, "
-                f"cameraAPI={repr(self.cameraAPI)})")
-
-    def __str__(self):
-        return self.description()
-
-    @property
-    def index(self):
-        """Camera index (`int`). This is the enumerated index of this camera.
-        """
-        return self._index
-
-    @index.setter
-    def index(self, value):
-        self._index = int(value)
-
-    @property
-    def name(self):
-        """Camera name (`str`). This is the camera name retrieved by the OS.
-        """
-        return self._name
-
-    @name.setter
-    def name(self, value):
-        self._name = str(value)
-
-    @property
-    def frameSize(self):
-        """Resolution (w, h) in pixels (`ArrayLike`).
-        """
-        return self._frameSize
-
-    @frameSize.setter
-    def frameSize(self, value):
-        assert len(value) == 2, "Value for `frameSize` must have length 2."
-        assert all([isinstance(i, int) for i in value]), (
-            "Values for `frameSize` must be integers.")
-
-        self._frameSize = value
-
-    @property
-    def frameRate(self):
-        """Resolution (min, max) in pixels (`ArrayLike`).
-        """
-        return self._frameRate
-
-    @frameRate.setter
-    def frameRate(self, value):
-        # assert len(value) == 2, "Value for `frameRateRange` must have length 2."
-        # assert all([isinstance(i, int) for i in value]), (
-        #     "Values for `frameRateRange` must be integers.")
-        # assert value[0] <= value[1], (
-        #     "Value for `frameRateRange` must be `min` <= `max`.")
-
-        self._frameRate = value
-
-    @property
-    def pixelFormat(self):
-        """Video pixel format (`str`). An empty string indicates this field is
-        not initialized.
-        """
-        return self._pixelFormat
-
-    @pixelFormat.setter
-    def pixelFormat(self, value):
-        self._pixelFormat = str(value)
-
-    @property
-    def codecFormat(self):
-        """Codec format, may be used instead of `pixelFormat` for some
-        configurations. Default is `''`.
-        """
-        return self._codecFormat
-
-    @codecFormat.setter
-    def codecFormat(self, value):
-        self._codecFormat = str(value)
-
-    @property
-    def cameraLib(self):
-        """Camera library these settings are targeted towards (`str`).
-        """
-        return self._cameraLib
-
-    @cameraLib.setter
-    def cameraLib(self, value):
-        self._cameraLib = str(value)
-
-    @property
-    def cameraAPI(self):
-        """Camera API in use to obtain this information (`str`).
-        """
-        return self._cameraAPI
-
-    @cameraAPI.setter
-    def cameraAPI(self, value):
-        self._cameraAPI = str(value)
-
-    def frameSizeAsFormattedString(self):
-        """Get image size as as formatted string.
-
-        Returns
-        -------
-        str
-            Size formatted as `'WxH'` (e.g. `'480x320'`).
-
-        """
-        return '{width}x{height}'.format(
-            width=self.frameSize[0],
-            height=self.frameSize[1])
-
-    def description(self):
-        """Get a description as a string.
-
-        Returns
-        -------
-        str
-            Description of the camera format as a human readable string.
-
-        """
-        codecFormat = self._codecFormat
-        pixelFormat = self._pixelFormat
-        codec = codecFormat if not pixelFormat else pixelFormat
-
-        return "[{name}] {width}x{height}@{frameRate}fps, {codec}".format(
-            #index=self.index,
-            name=self.name,
-            width=str(self.frameSize[0]),
-            height=str(self.frameSize[1]),
-            frameRate=str(self.frameRate),
-            codec=codec
-        )
-
-
-class StreamStatus:
-    """Descriptor class for stream status.
-
-    This class is used to report the current status of the stream read/writer.
-
-    Parameters
-    ----------
-    status : int
-        Status flag for the stream.
-    streamTime : float
-        Current stream time in seconds. This value increases monotonically and
-        is common to all webcams attached to the system.
-    recTime : float
-        If recording, this field will report the current timestamp within the
-        output file. Otherwise, this value is zero.
-    recBytes : float
-        If recording, this value indicates the number of bytes that have been
-        written out to file.
-
-    """
-    __slots__ = ['_status',
-                 '_streamTime',
-                 '_recTime',
-                 '_recBytes']
-
-    def __init__(self,
-                 status=NOT_STARTED,
-                 streamTime=0.0,
-                 recTime=0.0,
-                 recBytes=0):
-
-        self._status = int(status)
-        self._streamTime = float(streamTime)
-        self._recTime = float(recTime)
-        self._recBytes = int(recBytes)
-
-    @property
-    def status(self):
-        """Status flag for the stream (`int`).
-        """
-        return self._status
-
-    @property
-    def streamTime(self):
-        """Current stream time in seconds (`float`).
-
-        This value increases monotonically and is common timebase for all
-        cameras attached to the system.
-        """
-        return self._streamTime
-
-    @property
-    def recBytes(self):
-        """Current recording size on disk (`int`).
-
-        If recording, this value indicates the number of bytes that have been
-        written out to file.
-        """
-        return self._recBytes
-
-    @property
-    def recTime(self):
-        """Current recording time (`float`).
-
-        If recording, this field will report the current timestamp within the
-        output file. Otherwise, this value is zero.
-        """
-        return self._recTime
-
-
-class StreamData:
-    """Descriptor for camera stream data.
-
-    Instances of this class are produced by the stream reader/writer thread
-    which contain: metadata about the stream, frame image data (i.e. pixel
-    values), and the stream status.
-
-    Parameters
-    ----------
-    metadata : MovieMetadata
-        Stream metadata.
-    frameImage : object
-        Video frame image data.
-    streamStatus : StreamStatus
-        Video stream status.
-    cameraLib : str
-        Camera library in use to process the stream.
-
-    """
-    __slots__ = ['_metadata',
-                 '_frameImage',
-                 '_streamStatus',
-                 '_cameraLib']
-
-    def __init__(self, metadata, frameImage, streamStatus, cameraLib):
-        self._metadata = metadata
-        self._frameImage = frameImage
-        self._streamStatus = streamStatus
-        self._cameraLib = cameraLib
-
-    @property
-    def metadata(self):
-        """Stream metadata at the time the video frame was acquired
-        (`MovieMetadata`).
-        """
-        return self._metadata
-
-    @metadata.setter
-    def metadata(self, value):
-        if not isinstance(value, MovieMetadata) or value is not None:
-            raise TypeError("Incorrect type for property `metadata`, expected "
-                            "`MovieMetadata` or `None`.")
-
-        self._metadata = value
-
-    @property
-    def frameImage(self):
-        """Frame image data from the codec (`ffpyplayer.pic.Image`).
-        """
-        return self._frameImage
-
-    @frameImage.setter
-    def frameImage(self, value):
-        self._frameImage = value
-
-    @property
-    def streamStatus(self):
-        """Stream status (`StreamStatus`).
-        """
-        return self._streamStatus
-
-    @streamStatus.setter
-    def streamStatus(self, value):
-        if not isinstance(value, StreamStatus) or value is not None:
-            raise TypeError("Incorrect type for property `streamStatus`, "
-                            "expected `StreamStatus` or `None`.")
-
-        self._streamStatus = value
-
-    @property
-    def cameraLib(self):
-        """Camera library in use to obtain the stream (`str`). Value is
-        blank if `metadata` is `None`.
-        """
-        if self._metadata is not None:
-            return self._metadata.movieLib
-
-        return u''
-
-
-class StreamWriterThread(threading.Thread):
-    """Class for high-performance writing of video frames to disk asynchronously
-    using threading.
-
-    This thread is spawned along with a :class:`~psychopy.hardware.Camera`
-    instance, providing the capability to save real-time (live) video to disk
-    for later viewing or use. Communication and control of the thread are
-    done either by calling methods associated with this class, or directly
-     putting commands into the command queue, from another thread.
-
-    """
-    def __init__(self, mic=None):
-        threading.Thread.__init__(self)
-        self.daemon = True
-
-        self._mic = mic
-        self._commandQueue = queue.Queue()
-        self._writerClosedEvent = threading.Event()
-        self._writerClosedEvent.clear()
-
-        self._warmUpLock = threading.Lock()
-
-    def run(self):
-        """Main loop for the writer thread.
-
-        This waits on commands from the command queue and processes them
-        continuously until the `'end'` command is received. This thread may be
-        created along with the camera instance and killed when done.
-
-        If the stream format changes at any point, you should close the writer
-        and open a new one with updated parameters before writing more frames.
-
-        """
-        self._warmUpLock.acquire(blocking=False)
-        writer = None   # instance for the writer
-        filepath = ''  # path to the file
-        alive = True
-        while alive:
-            # block main thread until we are in the command loop
-            if self._warmUpLock.locked():
-                self._warmUpLock.release()
-
-            # process input commands
-            cmdOptCode, cmdVals = self._commandQueue.get(block=True)
-            if cmdOptCode == 'open':
-                # Open a file to write values to.
-                if writer is not None:
-                    raise IOError(
-                        "Attempted to start a new `MediaWriter` instance "
-                        "without closing the existing one first."
-                    )
-                filepath, writerOpts = cmdVals
-                # create a new writer instance
-                writer = MediaWriter(filepath, writerOpts)
-                self._commandQueue.task_done()
-            elif cmdOptCode == 'write_frame':  # write a frame
-                # Write a frame out to the file. Passing a boolean as arg[3]
-                # will tell the code whether to block until the frame has
-                # been written (or buffered and waiting) by the writer. Use
-                # `True` for synchronous operation and `False` for
-                # asynchronous.
-                if writer is None:
-                    raise IOError(
-                        'Got `write_frame` command but the writer has not '
-                        'been opened yet.')
-
-                colorData, pts, blockUntilDone = cmdVals
-                if not blockUntilDone:
-                    self._commandQueue.task_done()
-
-                frameWidth, frameHeight = colorData.get_size()
-                sws = SWScale(frameWidth, frameHeight,
-                    colorData.get_pixel_format(),
-                    ofmt='yuv420p')
-
-                # write the frame to the file
-                recordingBytes = writer.write_frame(
-                    img=sws.scale(colorData),
-                    pts=pts,
-                    stream=0)
-
-                logging.debug(
-                    'Writing {} bytes to file `{}`'.format(
-                        recordingBytes, filepath)
-                )
-
-                if blockUntilDone:
-                    self._commandQueue.task_done()
-            elif cmdOptCode == 'close':
-                # Close the file we are writing to but keep the writer
-                # thread hot. This allows for successive recordings to be
-                # made as quickly as possible without needing to spawn
-                # another thread each time.
-                if writer is None:
-                    raise IOError(
-                        "Attempted to close the `MediaWriter` instance "
-                        "without opening on first.")
-                writer.close()
-                writer = None
-                self._commandQueue.task_done()
-            elif cmdOptCode == 'end':  # end the thread
-                alive = False
-                continue
-
-        # if we have an open file, close it just in case
-        if writer is not None:
-            writer.close()
-
-        # set when the writer exits
-        self._commandQueue.task_done()  # when end is called
-
-        logging.debug('Media writer thread has been killed.')
-
-    @property
-    def commandQueue(self):
-        """The command queue for this thread (`queue.Queue`).
-        """
-        return self._commandQueue
-
-    def begin(self):
-        """Begin the file writer thread. Blocks until we can start accepting
-        commands.
-        """
-        self.start()
-        self._warmUpLock.acquire(blocking=True)
-
-    def end(self):
-        """Shutdown the thread.
-        """
-        self.sendCommand('end', (None,))  # blocks until done
-
-    def open(self, filePath, writerOpts):
-        """Open a file to writer frames to.
-
-        Parameters
-        ----------
-        filePath : str
-            Path to file to write frames to. This is usually a temporary
-            directory.
-        writerOpts : dict
-            Optional settings for the writer.
-
-        """
-        self.sendCommand('open', (filePath, writerOpts))  # blocks until done
-
-    def writeFrame(self, colorData, pts, blockUntilDone=False):
-        """Write a frame to the presently opened file.
-
-        Parameters
-        ----------
-        colorData : object
-            Image data to pass to the encoder.
-        pts : float
-            Presentation time stamp for the frame.
-        blockUntilDone : bool
-            Block this function until the frame has been written to disk.
-            Otherwise, this function will return immediately and the frame will
-            be written out asynchronously.
-
-        """
-        self.sendCommand('write_frame', (colorData, pts, blockUntilDone))
-
-    def close(self):
-        """Close the file. This will write out the result.
-        """
-        self.sendCommand('close', (None,))
-
-    def sendCommand(self, opcode, args):
-        """Send a command to this thread.
-
-        Parameters
-        ----------
-        opcode : str
-            Command key or operation code.
-        args : tuple
-            Arguments for the command.
-
-        """
-        self._commandQueue.put((opcode, args))
-        self._commandQueue.join()  # block until tasks are done
-
 
 class MovieStreamIOThread(threading.Thread):
     """Class for reading and writing streams asynchronously.
@@ -2132,264 +1587,6 @@ class Camera:
 # ------------------------------------------------------------------------------
 # Functions
 #
-
-def _getCameraInfoMacOS():
-    """Get a list of capabilities for the specified associated with a camera
-    attached to the system.
-
-    This is used by `getCameraInfo()` for querying camera details on MacOS.
-    Don't call this function directly unless testing.
-
-    Returns
-    -------
-    list of CameraInfo
-        List of camera descriptors.
-
-    """
-    if platform.system() != 'Darwin':
-        raise OSError(
-            "Cannot query cameras with this function, platform not 'Darwin'.")
-
-    # import objc  # may be needed in the future for more advanced stuff
-    import AVFoundation as avf  # only works on MacOS
-    import CoreMedia as cm
-
-    # get a list of capture devices
-    allDevices = avf.AVCaptureDevice.devices()
-
-    # get video devices
-    videoDevices = {}
-    devIdx = 0
-    for device in allDevices:
-        devFormats = device.formats()
-        if devFormats[0].mediaType() != 'vide':  # not a video device
-            continue
-
-        # camera details
-        cameraName = device.localizedName()
-
-        # found video formats
-        supportedFormats = []
-        for _format in devFormats:
-            # get the format description object
-            formatDesc = _format.formatDescription()
-
-            # get dimensions in pixels of the video format
-            dimensions = cm.CMVideoFormatDescriptionGetDimensions(formatDesc)
-            frameHeight = dimensions.height
-            frameWidth = dimensions.width
-
-            # Extract the codec in use, pretty useless since FFMPEG uses its
-            # own conventions, we'll need to map these ourselves to those
-            # values
-            codecType = cm.CMFormatDescriptionGetMediaSubType(formatDesc)
-
-            # Convert codec code to a FourCC code using the following byte
-            # operations.
-            #
-            # fourCC = ((codecCode >> 24) & 0xff,
-            #           (codecCode >> 16) & 0xff,
-            #           (codecCode >> 8) & 0xff,
-            #           codecCode & 0xff)
-            #
-            pixelFormat4CC = ''.join(
-                [chr((codecType >> bits) & 0xff) for bits in (24, 16, 8, 0)])
-
-            # Get the range of supported framerate, use the largest since the
-            # ranges are rarely variable within a format.
-            frameRateRange = _format.videoSupportedFrameRateRanges()[0]
-            frameRateMax = frameRateRange.maxFrameRate()
-            # frameRateMin = frameRateRange.minFrameRate()  # don't use for now
-
-            # Create a new camera descriptor
-            thisCamInfo = CameraInfo(
-                index=devIdx,
-                name=cameraName,
-                pixelFormat=pixelFormat4CC,  # macs only use pixel format
-                codecFormat=CAMERA_NULL_VALUE,
-                frameSize=(int(frameWidth), int(frameHeight)),
-                frameRate=frameRateMax,
-                cameraAPI=CAMERA_API_AVFOUNDATION
-            )
-
-            supportedFormats.append(thisCamInfo)
-
-            devIdx += 1
-
-        # add to output dictionary
-        videoDevices[cameraName] = supportedFormats
-
-    return videoDevices
-
-
-def _getCameraInfoWindows():
-    """Get a list of capabilities for the specified associated with a camera
-    attached to the system.
-
-    This is used by `getCameraInfo()` for querying camera details on Windows.
-    Don't call this function directly unless testing.
-
-    Returns
-    -------
-    list of CameraInfo
-        List of camera descriptors.
-
-    """
-    if platform.system() != 'Windows':
-        raise OSError(
-            "Cannot query cameras with this function, platform not 'Windows'.")
-
-    # FFPyPlayer can query the OS via DirectShow for Windows cameras
-    videoDevs, _, names = list_dshow_devices()
-
-    # get all the supported modes for the camera
-    videoDevices = {}
-
-    # iterate over names
-    devIndex = 0
-    for devURI in videoDevs.keys():
-        supportedFormats = []
-        cameraName = names[devURI]
-        for _format in videoDevs[devURI]:
-            pixelFormat, codecFormat, frameSize, frameRateRng = _format
-            _, frameRateMax = frameRateRng
-            temp = CameraInfo(
-                index=devIndex,
-                name=cameraName,
-                pixelFormat=pixelFormat,
-                codecFormat=codecFormat,
-                frameSize=frameSize,
-                frameRate=frameRateMax,
-                cameraAPI=CAMERA_API_DIRECTSHOW
-            )
-            supportedFormats.append(temp)
-            devIndex += 1
-
-        videoDevices[names[devURI]] = supportedFormats
-
-    return videoDevices
-
-
-# Mapping for platform specific camera getter functions used by `getCameras`.
-_cameraGetterFuncTbl = {
-    'Darwin': _getCameraInfoMacOS,
-    'Windows': _getCameraInfoWindows
-}
-
-
-def getCameras():
-    """Get information about installed cameras and their formats on this system.
-
-    Use `getCameraDescriptions` to get a mapping or list of human-readable
-    camera formats.
-
-    Returns
-    -------
-    dict
-        Mapping where camera names (`str`) are keys and values are and array of
-        `CameraInfo` objects.
-
-    """
-    systemName = platform.system()  # get the system name
-
-    # lookup the function for the given platform
-    getCamerasFunc = _cameraGetterFuncTbl.get(systemName, None)
-    if getCamerasFunc is None:  # if unsupported
-        raise OSError(
-            "Cannot get cameras, unsupported platform '{}'.".format(
-                systemName))
-
-    return getCamerasFunc()
-
-
-def getCameraDescriptions(collapse=False):
-    """Get a mapping or list of camera descriptions.
-
-    Camera descriptions are a compact way of representing camera settings and
-    formats. Description strings can be used to specify which camera device and
-    format to use with it to the `Camera` class.
-
-    Descriptions have the following format (example)::
-
-        '[Live! Cam Sync 1080p] 160x120@30fps, mjpeg'
-
-    This shows a specific camera format for the 'Live! Cam Sync 1080p' webcam
-    which supports 160x120 frame size at 30 frames per second. The last value
-    is the codec or pixel format used to decode the stream. Different pixel
-    formats and codecs vary in performance.
-
-    Parameters
-    ----------
-    collapse : bool
-        Return camera information as string descriptions instead of `CameraInfo`
-        objects. This provides a more compact way of representing camera formats
-        in a (reasonably) human-readable format.
-
-    Returns
-    -------
-    dict or list
-        Mapping (`dict`) of camera descriptions, where keys are camera names
-        (`str`) and values are a `list` of format description strings associated
-        with the camera. If `collapse=True`, all descriptions will be returned
-        in a single flat list. This might be more useful for specifying camera
-        formats from a single GUI list control.
-
-    """
-    connectedCameras = getCameras()
-
-    cameraDescriptions = {}
-    for devName, formats in connectedCameras.items():
-        cameraDescriptions[devName] = [
-            _format.description() for _format in formats]
-
-    if not collapse:
-        return cameraDescriptions
-
-    # collapse to a list if requested
-    collapsedList = []
-    for _, formatDescs in cameraDescriptions.items():
-        collapsedList.extend(formatDescs)
-
-    return collapsedList
-
-
-def renderVideo(outputFile, videoFile, audioFile=None):
-    """Render a video.
-
-    Combine visual and audio streams into a single movie file. This is used
-    mainly for compositing video and audio data for the camera. Video and audio
-    should have roughly the same duration.
-
-    Parameters
-    ----------
-    outputFile : str
-        Filename to write the movie to. Should have the extension of the file
-        too.
-    videoFile : str
-        Video file path.
-    audioFile : str or None
-        Audio file path. If not provided the movie file will simply be copied
-        to `outFile`.
-
-    Returns
-    -------
-    int
-        Size of the resulting file in bytes.
-
-    """
-    # merge audio and video tracks, we use MoviePy for this
-    videoClip = VideoFileClip(videoFile)
-
-    # if we have a microphone, merge the audio track in
-    if audioFile is not None:
-        audioClip = AudioFileClip(audioFile)
-        # add audio track to the video
-        videoClip.audio = CompositeAudioClip([audioClip])
-
-    # transcode with the format the user wants
-    videoClip.write_videofile(outputFile, verbose=False, logger=None)
-
-    return os.path.getsize(outputFile)
 
 
 if __name__ == "__main__":
