+++ psychopy/hardware/camera/constants.py
@@ -1,0 +1,58 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""Constants used by the camera module.
+"""
+
+# Part of the PsychoPy library
+# Copyright (C) 2002-2018 Jonathan Peirce (C) 2019-2022 Open Science Tools Ltd.
+# Distributed under the terms of the GNU General Public License (GPL).
+
+__all__ = [
+    'VIDEO_DEVICE_ROOT_LINUX',
+    'CAMERA_UNKNOWN_VALUE',
+    'CAMERA_NULL_VALUE',
+    # 'CAMERA_MODE_VIDEO',
+    # 'CAMERA_MODE_CV',
+    # 'CAMERA_MODE_PHOTO',
+    'CAMERA_TEMP_FILE_VIDEO',
+    'CAMERA_TEMP_FILE_AUDIO',
+    'CAMERA_API_AVFOUNDATION',
+    'CAMERA_API_DIRECTSHOW',
+    # 'CAMERA_API_VIDEO4LINUX',
+    # 'CAMERA_API_OPENCV',
+    'CAMERA_API_UNKNOWN',
+    'CAMERA_API_NULL',
+    'CAMERA_FRAMERATE_NTSC',
+    'CAMERA_FRAMERATE_NOMINAL_NTSC']
+
+VIDEO_DEVICE_ROOT_LINUX = '/dev'
+CAMERA_UNKNOWN_VALUE = u'Unknown'  # nominal fields where we don't have a value
+CAMERA_NULL_VALUE = u'Null'  # fields where we couldn't get a value
+# camera operating modes
+# CAMERA_MODE_VIDEO = u'video'
+# CAMERA_MODE_CV = u'cv'
+# CAMERA_MODE_PHOTO = u'photo'
+# default names for video and audio tracks in the temp directory
+CAMERA_TEMP_FILE_VIDEO = u'video.mp4'
+CAMERA_TEMP_FILE_AUDIO = u'audio.wav'
+
+# camera API flags, these specify which API camera settings were queried with
+CAMERA_API_AVFOUNDATION = u'AVFoundation'  # mac
+CAMERA_API_DIRECTSHOW = u'DirectShow'      # windows
+# CAMERA_API_VIDEO4LINUX = u'Video4Linux'  # linux
+# CAMERA_API_OPENCV = u'OpenCV'            # opencv, cross-platform API
+CAMERA_API_UNKNOWN = u'Unknown'            # unknown API
+CAMERA_API_NULL = u'Null'                  # empty field
+
+# camera libraries for playback nad recording
+CAMERA_LIB_FFPYPLAYER = u'FFPyPlayer'
+CAMERA_LIB_UNKNOWN = u'Unknown'
+CAMERA_LIB_NULL = u'Null'
+
+# special values
+CAMERA_FRAMERATE_NOMINAL_NTSC = '30.000030'
+CAMERA_FRAMERATE_NTSC = 30.000030
+
+
+if __name__ == "__main__":
+    pass
